Display all issues on map view (clickable to messages#index. Map dropdown in issues#new fixed to display only relevant maps
