+++ app/assets/stylesheets/pages/_map.scss
@@ -23,3 +23,7 @@
 .blue {
   background-color: lightblue;
 }
+
+.text-pink {
+  color: $pink;
+}
