+++ app/views/maps/_map_all_issues.html.erb
@@ -3,6 +3,7 @@
     <%= image_tag("https://upload.wikimedia.org/wikipedia/commons/9/9a/Sample_Floorplan.jpg") %>
   <% else %>
     <div class="map-image container" style="background-image: url('<%= cl_image_path @map.photo, height: 600, width: 600, crop: :pad %>')">
+      <div style="display: none" data-issues='<%= @issues.to_json %>'></div>
       <% for i in 1..30 do %>
         <div class="row">
         <% for j in 1..30 do %>
@@ -11,12 +12,5 @@
         </div>
       <% end %>
     </div>
-  <% end %>
-</div>
-<div class="btn-container">
-  <%= simple_form_for [@map.project, @issue] do |f| %>
-  <%= f.hidden_field :x_coordinate, value: 'x' %>
-  <%= f.hidden_field :y_coordinate, value: 'y' %>
-  <%= f.button :submit, value: 'Continue', class: 'create-btn emerald' %>
   <% end %>
 </div>
